requirements.txt
Flask
requests
python-dotenv
google-auth
google-auth-oauthlib
vertexai
